FINALLY GOT THIS CRAP TO WORK WOOP WOOP. All of this is only a problem because the screens and the GameScreenManager is instanced in a different document and scope than where we have to access the GSM-class's methods. If it where all in the same scope, we could just have a global variable the holds an object of the GSM-class. This variable can then be referenced inside the screen-classes without issue.
Instead of this we made all of our screens or scenes or whatever children of a larger GameScreenManager class. Functions in the GSM-class are responsible for getting and setting the currently active screen so that it can be drawn, activated and etc. It does this by having variables in the GSM-class that hold the objects of the screen-classes and then changes them accordingly. Because of this we have to create the screen objects first and then update the GSM-class's variables with a list (or in this case a dictionary for easy access) with them. But we can't just update the variables for the one instance of the GSM-class. If we do this the screen-objects, children of the GSM-class, won't have updated values for these variables and won't be able to access them when changing screens. It's because of this that the GSM-class's variables can't be defined in a constructor - this makes them object-specific and updates to them won't be visible for previously created objects such as the screens.

The buttons run code by passing a lambda function as a parameter for the constructor of the button. This lambda function can wrap another function with parameters that can then be run on the click of the button. As a result of this we can't directly set variables in the lambda functions and an update function was made to directly change the currentScreen variable of the GSM-class. This is not actually needed since we had made a crucial oversight - in the set_screen() function of the GSM-class we set a new value for self.currentState. This means that currentState is only affected in the respective object in which it was changed. Instead we can change the variable to be GameScreenManager.currentScreen, which is shared amongst all objects.