The previously mentioned issue with the update-function has been fixed.
Also I noticed that a class isn't even needed for this. Instead we could just remove the GSM-class and have all its variables and methods be global to the entire script. I like to thing that using this class could be a better practice since its variables and methods now won't be accessible for anything unrelated in the same script. We also just inherently now know when were performing an action that has anything with the screens to do.